Conflicts:
	xchange-mtgox/src/main/java/com/xeiam/xchange/mtgox/v1/service/account/MtGoxPollingAccountService.java
	xchange-mtgox/src/main/java/com/xeiam/xchange/mtgox/v1/service/trade/polling/MtGoxPollingTradeService.java

+++ xchange-mtgox/src/test/java/com/xeiam/xchange/mtgox/v1/service/marketdata/TickerJSONTest.java
@@ -50,13 +50,14 @@ public class TickerJSONTest {
     MtGoxTicker mtGoxTicker = mapper.readValue(is, MtGoxTicker.class);
 
     // Verify that the example data was unmarshalled correctly
-    assertThat("Unexpected Return Buy value", mtGoxTicker.getBuy().getValue(), equalTo(4.89002));
-    assertThat("Unexpected Return Last value", mtGoxTicker.getLast().getValue(), equalTo(4.89000));
-    assertThat("Unexpected Return Bid value", mtGoxTicker.getBuy().getValue(), equalTo(4.89002));
-    assertThat("Unexpected Return Ask value", mtGoxTicker.getSell().getValue(), equalTo(4.91227));
-    assertThat("Unexpected Return High value", mtGoxTicker.getHigh().getValue(), equalTo(4.98000));
-    assertThat("Unexpected Return Low value", mtGoxTicker.getLow().getValue(), equalTo(4.84272));
-    assertThat("Unexpected Return Volume value", mtGoxTicker.getVol().getValue(), equalTo(57759.66891627));
+    // TODO should use BigDecimal equality instead of double equality
+    assertThat("Unexpected Return Buy value", mtGoxTicker.getBuy().getValue().doubleValue(), equalTo(4.89002));
+    assertThat("Unexpected Return Last value", mtGoxTicker.getLast().getValue().doubleValue(), equalTo(4.89000));
+    assertThat("Unexpected Return Bid value", mtGoxTicker.getBuy().getValue().doubleValue(), equalTo(4.89002));
+    assertThat("Unexpected Return Ask value", mtGoxTicker.getSell().getValue().doubleValue(), equalTo(4.91227));
+    assertThat("Unexpected Return High value", mtGoxTicker.getHigh().getValue().doubleValue(), equalTo(4.98000));
+    assertThat("Unexpected Return Low value", mtGoxTicker.getLow().getValue().doubleValue(), equalTo(4.84272));
+    assertThat("Unexpected Return Volume value", mtGoxTicker.getVol().getValue().doubleValue(), equalTo(57759.66891627));
   }
 
   @Test
@@ -76,13 +77,14 @@ public class TickerJSONTest {
     MtGoxTicker mtGoxTicker = mapper.readValue(mapper.writeValueAsString(userInMap.get("ticker")), MtGoxTicker.class);
 
     // Verify that the example data was unmarshalled correctly
-    assertThat("Unexpected Return Buy value", mtGoxTicker.getBuy().getValue(), equalTo(5.10991));
-    assertThat("Unexpected Return Last value", mtGoxTicker.getLast().getValue(), equalTo(5.10991));
-    assertThat("Unexpected Return Bid value", mtGoxTicker.getBuy().getValue(), equalTo(5.10991));
-    assertThat("Unexpected Return Ask value", mtGoxTicker.getSell().getValue(), equalTo(5.11000));
-    assertThat("Unexpected Return High value", mtGoxTicker.getHigh().getValue(), equalTo(5.12500));
-    assertThat("Unexpected Return Low value", mtGoxTicker.getLow().getValue(), equalTo(5.07000));
-    assertThat("Unexpected Return Volume value", mtGoxTicker.getVol().getValue(), equalTo(15475.00497509));
+    // TODO should use BigDecimal equality instead of double equality
+    assertThat("Unexpected Return Buy value", mtGoxTicker.getBuy().getValue().doubleValue(), equalTo(5.10991));
+    assertThat("Unexpected Return Last value", mtGoxTicker.getLast().getValue().doubleValue(), equalTo(5.10991));
+    assertThat("Unexpected Return Bid value", mtGoxTicker.getBuy().getValue().doubleValue(), equalTo(5.10991));
+    assertThat("Unexpected Return Ask value", mtGoxTicker.getSell().getValue().doubleValue(), equalTo(5.11000));
+    assertThat("Unexpected Return High value", mtGoxTicker.getHigh().getValue().doubleValue(), equalTo(5.12500));
+    assertThat("Unexpected Return Low value", mtGoxTicker.getLow().getValue().doubleValue(), equalTo(5.07000));
+    assertThat("Unexpected Return Volume value", mtGoxTicker.getVol().getValue().doubleValue(), equalTo(15475.00497509));
 
   }
 }
